Organizing and Error Handling
Organizing code by splitting it into 3 files. Endpoints/Flask things remain in app, database including classes is in data, and code for testing is in testing.

Added error handling at recipe submission to catch malicious submissions.
